The initial deal diagram generator
First public version of deal diagram generator is used for my personal
blog. There is still a lot of work to do in styles, animations, visualising
playing errors, code comments and documentation. But current version is
enough to provide minimal functionality.

The javascript side depend heavily on jquery library that has to be
loaded before loading the deal.js.

I won't prevent anyone using this version for their blog but it
may help to talk me first to avoid issues.

Signed-off-by: Pauli Nieminen <suokkos@gmail.com>